capstone_dog_app/requirements/requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mkl-fft==1.2.0
mkl-random==1.1.1
numpy
olefile==0.46
Pillow
six
torch
torchvision
typing-extensions
Werkzeug==1.0.1
wincertstore==0.2
